Debating adding in a sigmoid function soon

+++ MyGame.cs
@@ -9,8 +9,8 @@ namespace LearnerAgent
     public class MyGame : Game
     {
         // Temporary variables for testing
-        public static int X = 0;
-        public static int Y = 0;
+        public static float X = 0;
+        public static float Y = 0;
         
         // Set the camera scale
         private float SPRITE_SCALE = 0.25f;
@@ -84,7 +84,7 @@ namespace LearnerAgent
                 null,
                 null,
                 transform);
-            _spriteBatch.Draw(_agent, new Rectangle(X, Y, 128, 128), Color.White);
+            _spriteBatch.Draw(_agent, new Rectangle((int)X, (int)Y, 128, 128), Color.White);
             _spriteBatch.End();
 
             base.Draw(gameTime);
